build: add `grub-theme-vimix-very-dark-blue.install`
"Pacman has the ability to store and execute a package-specific script
when it installs, removes, or upgrades a package. This allows a package
to configure itself after installation and perform an opposite action
upon removal." (https://man.archlinux.org/man/PKGBUILD.5)

Add the following scripts:

    post_install
    post_upgrade
    post_remove

This assures that `/boot/grub/grub.cfg` is up-to-date when a package
update happened. This also warns the user in case GRUB still uses this
theme after uninstalling it.