feat(install)!: improve status messages
Improve the clarity of the status messages printed during installation,
upgrade, and removal of the GRUB theme by providing more detailed
information and clearer instructions.

BREAKING CHANGE: The output messages for installation, upgrade, and
removal of the GRUB theme have been significantly improved, and may not
be compatible with tools that relied on the previous output format.
Please update any affected tools accordingly.